app/server/requirements.txt
pulsar-client==2.4.1
gevent
gunicorn
flask
influxdb
flask_cors
timeout-decorator